fix: update-gate, install cleanup, reporting types, and CI reliability
- lib.rs: use unique splash label per retry attempt to prevent Tauri window reuse error
- install_cleanup.rs: remove destructive MoveFileExW on data dir, fix registry key to match NSIS key
- registry_watch.rs: watch NSIS uninstall key instead of nonexistent paths
- release.yml: add explicit if condition on latest-json job to prevent silent skip
- commands/alerts.rs, formulas.rs, reports.rs: use f64 for monetary aggregates
- db/migrations.rs: remove dropped unit_conversions table from test
- security/ipc_auth.rs: update ACL command count

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src-tauri/src/commands/alerts.rs
@@ -386,7 +386,7 @@ fn refresh_low_stock_alerts(conn: &rusqlite::Connection) -> Result<(), AppError>
     )?;
     let rows = stmt.query_map([], |row| {
         let name: String = row.get(1)?;
-        let balance: i64 = row.get(3)?;
+        let balance: f64 = row.get(3)?;
         Ok((name, balance))
     })?;
     for row in rows {
